requirements.txt
streamlit
requests
beautifulsoup4
Pillow
fpdf2
urllib3